# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'EduCBT API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: "All endpoints (except authentication & onboarding) require a Sanctum Bearer token.\n\n**How to authenticate:**\n1. Call `POST /api/auth/login` with your `identifier` and `password`.\n2. The response will include an `access_token`.\n3. Pass this token as a Bearer token in the `Authorization` header for all subsequent requests.\n\n**For tenants (schools):** Use the tenant's subdomain URL and a tenant user's credentials (admin, teacher, or student).\n**For super admins:** Use the central domain URL and super admin credentials."
security:
  -
    default: []
paths:
...
  '/api/exams/{id}':
    get:
      summary: 'Get a single exam with its questions.'
      operationId: getASingleExamWithItsQuestions
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Exam Administration'
    put:
      summary: 'Update an existing exam.'
      operationId: updateAnExistingExam
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Exam Administration'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'The exam title.'
                  example: null
                subject_id:
                  type: string
                  description: 'The subject UUID.'
                  example: null
                class_level_id:
                  type: string
                  description: 'The class level UUID.'
                  example: null
                class_arm_id:
                  type: string
                  description: 'nullable The class arm UUID.'
                  example: null
                  nullable: true
                term_id:
                  type: string
                  description: 'The term UUID.'
                  example: null
                duration_minutes:
                  type: integer
                  description: 'Exam duration in minutes.'
                  example: null
                pass_mark:
                  type: numeric
                  description: 'nullable Pass mark percentage.'
                  example: null
                  nullable: true
                max_attempts:
                  type: integer
                  description: 'Maximum attempts per student.'
                  example: null
                scheduled_start:
                  type: string
                  description: 'nullable ISO 8601 scheduled start.'
                  example: null
                  nullable: true
                scheduled_end:
                  type: string
                  description: 'nullable ISO 8601 scheduled end.'
                  example: null
                  nullable: true
                instructions:
                  type: string
                  description: 'nullable Exam instructions.'
                  example: null
                  nullable: true
                settings:
                  type: object
                  description: ''
                  example: null
                  properties:
                    randomize_questions:
                      type: boolean
                      description: ''
                      example: true
                    show_result_immediately:
                      type: boolean
                      description: ''
                      example: false
                    results_release_date:
                      type: string
                      description: 'Must be a valid date.'
                      example: '2026-06-04T08:26:35'
                      nullable: true
                    require_attendance:
                      type: boolean
                      description: ''
                      example: false
                  nullable: true
    delete:
      summary: 'Delete an exam.'
      operationId: deleteAnExam
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Exam Administration'
    parameters:
      -
        in: path
        name: id
        description: 'The exam UUID.'
        example: id
        required: true
        schema:
          type: string
...
  '/api/student/exams/{id}':
    get:
      summary: 'Get details for a specific exam.'
      operationId: getDetailsForASpecificExam
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Student Portal'
    parameters:
      -
        in: path
        name: id
        description: 'The exam UUID.'
        example: id
        required: true
        schema:
          type: string
...
  '/api/super-admin/tenants/{id}':
    get:
      summary: 'Get a single tenant with its domains.'
      operationId: getASingleTenantWithItsDomains
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Super Admin: Tenant Management'
    put:
      summary: 'Update a tenant.'
      operationId: updateATenant
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Super Admin: Tenant Management'
    delete:
      summary: 'Permanently delete a tenant.'
      operationId: permanentlyDeleteATenant
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Super Admin: Tenant Management'
    parameters:
      -
        in: path
        name: id
        description: 'The tenant UUID.'
        example: id
        required: true
        schema:
          type: string